Used Pundit to ensure proper authorization for products/buyers/sellers/orders pages

+++ app/controllers/products_controller.rb
@@ -1,37 +1,53 @@
 class ProductsController < ApplicationController
   before_action :authenticate_account!
   before_action :set_product, only: [:show, :edit, :update, :destroy]
+  rescue_from ActiveRecord::RecordNotFound, with: :invalid_product
+  
 
+
+  def pundit_user
+    current_account
+  end
+  
   # GET /products
   # GET /products.json
   def index
-     if (params[:seller_id])
-      @seller = Seller.find(params[:seller_id])
-      @products = @seller.products
-     else
-      @products = Product.all
-     end   
+    #  if (params[:seller_id])
+    #   @seller = Seller.find(params[:seller_id])
+    #   @products = @seller.products
+    # else
+    #   @products = Product.all
+    # end   
+
+    # The following is much neater than above, isn't it?
+    authorize Product 
+    @products = policy_scope(Product)
   end
 
   # GET /products/1
   # GET /products/1.json
   def show
-    @orders = @product.orders
+    authorize @product 
+    @product.orders
+     
   end
 
   # GET /products/new
   def new
     @product = Product.new
+    authorize @product
   end
 
   # GET /products/1/edit
   def edit
+    authorize @product
   end
 
   # POST /products
   # POST /products.json
   def create
     @product = Product.new(product_params)
+    authorize @product
     if current_account && current_account.accountable_type == "Seller"
         @product.seller = current_account.accountable
     end
@@ -50,6 +66,7 @@ class ProductsController < ApplicationController
   # PATCH/PUT /products/1
   # PATCH/PUT /products/1.json
   def update
+    authorize @product
     respond_to do |format|
       if @product.update(product_params)
         format.html { redirect_to @product, notice: 'Product was successfully updated.' }
@@ -67,6 +84,8 @@ class ProductsController < ApplicationController
   # DELETE /products/1
   # DELETE /products/1.json
   def destroy
+    authorize @product
+    @product.destroy
     respond_to do |format|
       if @product.destroy
         format.html { redirect_to products_url, notice: 'Product was successfully destroyed.' }
@@ -88,5 +107,10 @@ class ProductsController < ApplicationController
     # Never trust parameters from the scary internet, only allow the white list through.
     def product_params
       params.require(:product).permit(:title, :description, :image_url, :price)
+    end
+    
+    def invalid_product
+      logger.error "Attempt to access invalid product #{params[:id]}"
+      redirect_to store_index_url, notice: 'Invalid product'
     end
 end
